requirements.txt
yfinance
pandas
groq
python-dotenv
schedule